requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq 
chromadb
sentence-transformers
langchain_huggingface
pypdf
langchain_chroma